modify exit button

+++ frontend/edgeinfo.py
@@ -1,4 +1,4 @@
-from PyQt5.QtWidgets import QWidget, QGridLayout, QLabel, QLineEdit, QPushButton
+from PyQt5.QtWidgets import QWidget, QGridLayout, QLabel, QLineEdit, QPushButton, QMessageBox
 from PyQt5.QtCore import Qt
 from frontend.vertexinfo import EditLabel
 from backend.edge import update_edge, delete_edges
@@ -59,8 +59,12 @@ class EdgeInfo(QWidget):
         delete_button.clicked.connect(lambda: self.delete_clicked())
 
     def delete_clicked(self):
-        delete_edges(self.parent.graph, self.edge)
-        self.parent.view.update_view()
+        reply = QMessageBox.question(self, '', 'Are you sure want to delete this edge?',
+                                     QMessageBox.Yes, QMessageBox.No)
+        if reply == QMessageBox.Yes:
+            delete_edges(self.parent.graph, self.edge)
+            self.parent.view.update_view()
+            self.parent.clear_layout(self.parent.info_layout)
 
     @staticmethod
     def text_edited(value, edit):
